refactor(design): carve the agent composer into modules under agent-chat/ (#8673)
