+++ packages/design/src/agent-chat/HarnessWidget.css
@@ -1,0 +1,22 @@
+.kp-agent-chat__widget {
+	display: flex;
+	flex-direction: column;
+	gap: var(--s-2);
+}
+
+.kp-agent-chat__widget-title {
+	margin: 0;
+	color: var(--text-secondary);
+	font: var(--t-meta);
+	font-weight: 700;
+}
+
+.kp-agent-chat__widget pre {
+	max-height: calc(var(--s-8) * 8);
+	margin: 0;
+	overflow: auto;
+	color: var(--text-primary);
+	font: var(--t-body);
+	white-space: pre-wrap;
+	word-break: break-word;
+}
