+++ packages/design/src/agent-chat/PiExtensionDialog.css
@@ -1,0 +1,5 @@
+.kp-agent-chat__extension-options {
+	display: flex;
+	flex-direction: column;
+	gap: var(--s-2);
+}
